Rename wrappers and components to containers and event components to component directory
