--- src/components/Authentication/index.js
@@ -1,5 +1,0 @@
-import Login from './Login'
-import Logout from './Logout'
-import LoginRedirect from './LoginRedirect'
-
-export { Login, Logout, LoginRedirect }
