+++ src/pages/EventPage.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { connect } from "react-redux";
 import { setEvent } from "../actions/PageActions";
-import EventFormWrapper from '../components/EventForm/EventFormWrapper';
+import EventFormContainer from '../components/EventFormContainer';
 import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
@@ -45,7 +45,7 @@ const EventPage = (props) => {
         return (
             <div className={classes.layout}>
                 <Paper className={classes.paper}>
-                    <EventFormWrapper event={event} />
+                    <EventFormContainer event={event} />
                 </Paper>
             </div>
         )
